Resolving the issue of no data reporting in the BOE environment due to the inability to find message serialization and deserialization classes
